HlsParser인터페이스 구현,
